LHY 0425 fin

+++ 00_SemiProject/WebContent/owner/board/reply_modify.jsp
@@ -36,7 +36,7 @@
 				<div class="button_div">
 					<input type="hidden" value="${dto.getInqho_no() }" name="no">
 					<input type="submit" value="수정">
-					<input type="button" value="삭제" onclick="location.href='<%=request.getContextPath() %>/owner_reply_delete.do?no=${dto.getInqho_no() }'">
+					<input type="button" value="삭제" onclick="if(confirm('정말로 삭제하시겠습니까.')){location.href='<%=request.getContextPath() %>/owner_reply_delete.do?no=${dto.getInqho_no() }'}">
 					<input type="button" value="목록" onclick="board_move()">
 				</div>
 			</form>
